cmd/bootnode: move bootnode to own package
