+++ app/app_test.go
@@ -39,7 +39,7 @@ import (
 	"github.com/obolnetwork/charon/app/log"
 	"github.com/obolnetwork/charon/app/version"
 	"github.com/obolnetwork/charon/cluster"
-	"github.com/obolnetwork/charon/cmd"
+	"github.com/obolnetwork/charon/cmd/bootnode"
 	"github.com/obolnetwork/charon/core"
 	"github.com/obolnetwork/charon/core/priority"
 	"github.com/obolnetwork/charon/p2p"
@@ -294,7 +294,7 @@ func startBootnode(ctx context.Context, t *testing.T) (string, <-chan error) {
 
 	errChan := make(chan error, 1)
 	go func() {
-		errChan <- cmd.RunBootnode(ctx, cmd.BootnodeConfig{
+		errChan <- bootnode.Run(ctx, bootnode.Config{
 			DataDir:  dir,
 			HTTPAddr: addr,
 			P2PConfig: p2p.Config{
